fix: allow custom notification messages building
